+++ lib/flack/app/helpers.rb
@@ -126,8 +126,8 @@ class Flack::App
 
     h = {}
 
-    h['self'] = {
-      href: rel(env, env['REQUEST_PATH']) }
+    h['self'] = { href: rel(env, env['REQUEST_PATH'] || env['PATH_INFO']) }
+
     m = env['REQUEST_METHOD']
     h['self'][:method] = m unless %w[ GET HEAD ].include?(m)